CohortGenerator-package {CohortGenerator}	R Documentation

CohortGenerator: An R Package for Cohort Generation Against the OMOP CDM

Description

An R package for that encapsulates the functions for generating cohorts against the OMOP CDM.

Author(s)

Maintainer: Anthony Sena sena@ohdsi.org

Authors:

Gowtham Rao gowthamrao@gmail.com

Martijn Schuemie schuemie@ohdsi.org

Other contributors:

Observational Health Data Science and Informatics [copyright holder]

See Also

Useful links:

https://ohdsi.github.io/CohortGenerator/

https://github.com/OHDSI/CohortGenerator

Report bugs at https://github.com/OHDSI/CohortGenerator/issues

[Package CohortGenerator version 0.4.0 Index]