Cohort Subset Operator

A subset of type cohort - subset a population to only those contained within defined cohort

Super class

CohortGenerator::SubsetOperator -> CohortSubsetOperator

Active bindings

cohortIds

Integer ids of cohorts to subset to

cohortCombinationOperator

How to combine the cohorts

negate

Inverse the subset rule? TRUE will take the patients NOT in the subset

windows

list of time windows to use when evaluating the subset cohort relative to the target cohort

Methods

Public methods

CohortSubsetOperator$new()

CohortSubsetOperator$toList()

CohortSubsetOperator$getAutoGeneratedName()

CohortSubsetOperator$clone()

Inherited methods

CohortGenerator::SubsetOperator$classname()

CohortGenerator::SubsetOperator$getQueryBuilder()

CohortGenerator::SubsetOperator$isEqualTo()

CohortGenerator::SubsetOperator$publicFields()

CohortGenerator::SubsetOperator$toJSON()

Method new()

Usage

CohortSubsetOperator$new(definition = NULL)

Arguments

definition

json character or list - definition of subset operator

Returns

instance of object to List

Method toList()

List representation of object Get auto generated name

Usage

CohortSubsetOperator$toList()

Method getAutoGeneratedName()

name generated from subset operation properties

Usage

CohortSubsetOperator$getAutoGeneratedName()

Returns

character

Method clone()

The objects of this class are cloneable with this method.

Usage

CohortSubsetOperator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.