Demographic Subset Operator

Operators for subsetting a cohort by demographic criteria

Value

char vector Get auto generated name

Super class

CohortGenerator::SubsetOperator -> DemographicSubsetOperator

Active bindings

ageMin

Int between 0 and 99999 - minimum age

ageMax

Int between 0 and 99999 - maximum age

gender

vector of gender concept IDs

race

character string denoting race

ethnicity

character string denoting ethnicity

Methods

Public methods

DemographicSubsetOperator$toList()

DemographicSubsetOperator$mapGenderConceptsToNames()

DemographicSubsetOperator$getAutoGeneratedName()

DemographicSubsetOperator$toJSON()

DemographicSubsetOperator$isEqualTo()

DemographicSubsetOperator$getGender()

DemographicSubsetOperator$getRace()

DemographicSubsetOperator$getEthnicity()

DemographicSubsetOperator$clone()

Inherited methods

CohortGenerator::SubsetOperator$classname()

CohortGenerator::SubsetOperator$getQueryBuilder()

CohortGenerator::SubsetOperator$initialize()

CohortGenerator::SubsetOperator$publicFields()

Method toList()

List representation of object Map gender concepts to names

Usage

DemographicSubsetOperator$toList()

Method mapGenderConceptsToNames()

Usage

DemographicSubsetOperator$mapGenderConceptsToNames(
  mapping = list(`8507` = "males", `8532` = "females", `0` = "unknown gender")
)

Arguments

mapping

optional list of mappings for concept id to nouns

Method getAutoGeneratedName()

name generated from subset operation properties

Usage

DemographicSubsetOperator$getAutoGeneratedName()

Returns

character

Method toJSON()

json serialized representation of object

Usage

DemographicSubsetOperator$toJSON()

Method isEqualTo()

Compare Subset to another

Usage

DemographicSubsetOperator$isEqualTo(criteria)

Arguments

criteria

DemographicSubsetOperator instance

Method getGender()

Gender getter - used when constructing SQL to default NULL to an empty string

Usage

DemographicSubsetOperator$getGender()

Method getRace()

Race getter - used when constructing SQL to default NULL to an empty string

Usage

DemographicSubsetOperator$getRace()

Method getEthnicity()

Ethnicity getter - used when constructing SQL to default NULL to an empty string

Usage

DemographicSubsetOperator$getEthnicity()

Method clone()

The objects of this class are cloneable with this method.

Usage

DemographicSubsetOperator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.