SubsetCohortWindow {CohortGenerator}	R Documentation

to List

Description

Representation of a time window to use when subsetting a target cohort with a subset cohort

Details

SubsetCohortWindow settings

Active bindings

startDay

Integer

endDay

Integer

targetAnchor

Boolean

Methods

Public methods

SubsetCohortWindow$toList()

SubsetCohortWindow$toJSON()

SubsetCohortWindow$isEqualTo()

SubsetCohortWindow$clone()

Method toList()

List representation of object To JSON

Usage

SubsetCohortWindow$toList()

Method toJSON()

json serialized representation of object Is Equal to

Usage

SubsetCohortWindow$toJSON()

Method isEqualTo()

Compare SubsetCohortWindow to another

Usage

SubsetCohortWindow$isEqualTo(criteria)

Arguments

criteria

SubsetCohortWindow instance

Method clone()

The objects of this class are cloneable with this method.

Usage

SubsetCohortWindow$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package CohortGenerator version 0.8.0 Index]