Abstract base class for subsets.

Abstract Base Class for subsets. Subsets should inherit from this and implement their own requirements.

See also

Active bindings

name

name of subset operation - should describe what the operation does e.g. "Males under the age of 18", "Exposed to Celecoxib"

Methods

Public methods

SubsetOperator$new()

SubsetOperator$classname()

SubsetOperator$getAutoGeneratedName()

SubsetOperator$getQueryBuilder()

SubsetOperator$publicFields()

SubsetOperator$isEqualTo()

SubsetOperator$toList()

SubsetOperator$toJSON()

SubsetOperator$clone()

Method new()

Usage

SubsetOperator$new(definition = NULL)

Arguments

definition

json character or list - definition of subset operator

Returns

instance of object Class Name

Method classname()

Class name of object Get auto generated name

Usage

SubsetOperator$classname()

Method getAutoGeneratedName()

Not intended to be used - should be implemented in subclasses Return query builder instance

Usage

SubsetOperator$getAutoGeneratedName()

Method getQueryBuilder()

Return query builder instance Public Fields

Usage

SubsetOperator$getQueryBuilder(id)

Arguments

id

- integer that should be unique in the sql (e.g. increment it by one for each subset operation in set)

Method publicFields()

Publicly settable fields of object Is Equal to

Usage

SubsetOperator$publicFields()

Method isEqualTo()

Compare Subsets - are they identical or not? Checks all fields and settings

Usage

SubsetOperator$isEqualTo(subsetOperatorB)

Arguments

subsetOperatorB

A subset to test equivalence to To list

Method toList()

convert to List representation To Json

Usage

SubsetOperator$toList()

Method toJSON()

convert to json serialized representation

Usage

SubsetOperator$toJSON()

Returns

list representation of object as json character

Method clone()

The objects of this class are cloneable with this method.

Usage

SubsetOperator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.