createCohortSubset {CohortGenerator}	R Documentation

A definition of subset functions to be applied to a set of cohorts

Description

A definition of subset functions to be applied to a set of cohorts

Usage

createCohortSubset(
  name = NULL,
  cohortIds,
  cohortCombinationOperator,
  negate,
  startWindow,
  endWindow
)

Arguments

name
optional name of operator

cohortIds
integer - set of cohort ids to subset to

cohortCombinationOperator
"any" or "all" if using more than one cohort id allow a subject to be in any cohort or require that they are in all cohorts in specified windows

negate
The opposite of this definition - include patients who do NOT meet the specified criteria (NOT YET IMPLEMENTED)

startWindow
A SubsetCohortWindow that patients must fall inside (see createSubsetCohortWindow)

endWindow
A SubsetCohortWindow that patients must fall inside (see createSubsetCohortWindow)

Value

a CohortSubsetOperator instance

[Package CohortGenerator version 0.8.0 Index]