generateCohort {CohortGenerator}	R Documentation

Generates a cohort

Description

This function is used by generateCohortSet to generate a cohort against the CDM.

Usage

generateCohort(
  cohortId = NULL,
  cohortDefinitionSet,
  connection = NULL,
  connectionDetails = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema,
  cohortDatabaseSchema,
  cohortTableNames,
  stopIfError = TRUE,
  incremental,
  recordKeepingFile
)

Arguments

cohortId
The cohortId in the list of cohortDefinitionSet

cohortDefinitionSet
The cohortDefinitionSet argument must be a data frame with the following columns:

cohortId

The unique integer identifier of the cohort

cohortName

The cohort's name

sql

The OHDSI-SQL used to generate the cohort

Optionally, this data frame may contain:

json

The Circe JSON representation of the cohort

connection
An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

cdmDatabaseSchema
Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

tempEmulationSchema
Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created.

cohortDatabaseSchema
Schema name where your cohort tables reside. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

cohortTableNames
The names of the cohort tables. See getCohortTableNames for more details.

stopIfError
When set to true, an error in processing will call the stop() command to notify the parent calling function that an error occurred.

incremental
Create only cohorts that haven't been created before?

recordKeepingFile
If incremental = TRUE, this file will contain information on cohorts already generated

[Package CohortGenerator version 0.3.0 Index]