getCohortTableNames {CohortGenerator}	R Documentation

Used to get a list of cohort table names to use when creating the cohort tables

Description

This function creates a list of table names used by createCohortTables to specify the table names to create. Use this function to specify the names of the main cohort table and cohort statistics tables.

Usage

getCohortTableNames(
  cohortTable = "cohort",
  cohortInclusionTable = paste0(cohortTable, "_inclusion"),
  cohortInclusionResultTable = paste0(cohortTable, "_inclusion_result"),
  cohortInclusionStatsTable = paste0(cohortTable, "_inclusion_stats"),
  cohortSummaryStatsTable = paste0(cohortTable, "_summary_stats"),
  cohortCensorStatsTable = paste0(cohortTable, "_censor_stats")
)

Arguments

cohortTable
Name of the cohort table.

cohortInclusionTable
Name of the inclusion table, one of the tables for storing inclusion rule statistics.

cohortInclusionResultTable
Name of the inclusion result table, one of the tables for storing inclusion rule statistics.

cohortInclusionStatsTable
Name of the inclusion stats table, one of the tables for storing inclusion rule statistics.

cohortSummaryStatsTable
Name of the summary stats table, one of the tables for storing inclusion rule statistics.

cohortCensorStatsTable
Name of the censor stats table, one of the tables for storing inclusion rule statistics.

Value

A list of the table names as specified in the parameters to this function.

[Package CohortGenerator version 0.3.0 Index]