getRequiredTasks {CohortGenerator}	R Documentation

Get a list of tasks required when running in incremental mode

Description

This function will attempt to check the recordKeepingFile to determine if a list of operations have completed by comparing the keys passed into the function with the checksum supplied

Usage

getRequiredTasks(..., checksum, recordKeepingFile)

Arguments

...
Parameter values used to identify the key in the incremental record keeping file

checksum
The checksum representing the operation to check

recordKeepingFile
A file path to a CSV file containing the record keeping information.

Value

Returns a list of outstanding tasks based on inspecting the full contents of the record keeping file

[Package CohortGenerator version 0.4.0 Index]