isCamelCase {CohortGenerator}	R Documentation

Used to check if a string is in lower camel case

Description

This function is used check if a string conforms to the lower camel case format.

Usage

isCamelCase(x)

Arguments

x
The string to evaluate

Value

TRUE if the string is in lower camel case

[Package CohortGenerator version 0.8.0 Index]