isSnakeCase {CohortGenerator}	R Documentation

Used to check if a string is in snake case

Description

This function is used check if a string conforms to the snake case format.

Usage

isSnakeCase(x)

Arguments

x
The string to evaluate

Value

TRUE if the string is in snake case

[Package CohortGenerator version 0.8.0 Index]