readCsv {CohortGenerator}	R Documentation

Used to read a .csv file

Description

This function is used to centralize the function for reading .csv files across the HADES ecosystem. This function will automatically convert from snake_case in the file to camelCase in the data.frame returned as is the standard described in: https://ohdsi.github.io/Hades/codeStyle.html#Interfacing_between_R_and_SQL

Usage

readCsv(file, warnOnCaseMismatch = TRUE)

Arguments

file
The .csv file to read.

warnOnCaseMismatch
When TRUE, raise a warning if column headings in the .csv are not in snake_case format

Value

A tibble with the .csv contents

[Package CohortGenerator version 0.8.0 Index]