recordTasksDone {CohortGenerator}	R Documentation

Record a task as complete

Description

This function will record a task as completed in the recordKeepingFile

Usage

recordTasksDone(..., checksum, recordKeepingFile, incremental = TRUE)

Arguments

...
Parameter values used to identify the key in the incremental record keeping file

checksum
The checksum representing the operation to check

recordKeepingFile
A file path to a CSV file containing the record keeping information.

incremental
When TRUE, this function will record tasks otherwise it will return without attempting to perform any action

[Package CohortGenerator version 0.3.0 Index]