Used in incremental mode to save values to a file

When running in incremental mode, we may need to update results in a CSV file. This function will replace the data in fileName based on the key parameters

Usage

saveIncremental(data, fileName, ...)

Arguments

data

The data to record in the file

fileName

A CSV holding results in the same structure as the data parameter

...

Parameter values used to identify the key in the results file