CohortSubsetOperator {CohortGenerator}	R Documentation

Cohort Subset Operator

Description

A subset of type cohort - subset a population to only those contained within defined cohort to List

Super class

CohortGenerator::SubsetOperator -> CohortSubsetOperator

Active bindings

cohortIds

Integer ids of cohorts to subset to

cohortCombinationOperator

How to combine the cohorts

negate

Inverse the subset rule? TRUE will take the patients NOT in the subset

startWindow

The time window to use evaluating the subset cohort start relative to the target cohort

endWindow

The time window to use evaluating the subset cohort end relative to the target cohort

Methods

Public methods

CohortSubsetOperator$toList()

CohortSubsetOperator$getAutoGeneratedName()

CohortSubsetOperator$clone()

Inherited methods

CohortGenerator::SubsetOperator$classname()

CohortGenerator::SubsetOperator$getQueryBuilder()

CohortGenerator::SubsetOperator$initialize()

CohortGenerator::SubsetOperator$isEqualTo()

CohortGenerator::SubsetOperator$publicFields()

CohortGenerator::SubsetOperator$toJSON()

Method toList()

List representation of object Get auto generated name

Usage

CohortSubsetOperator$toList()

Method getAutoGeneratedName()

name generated from subset operation properties

Usage

CohortSubsetOperator$getAutoGeneratedName()

Returns

character

Method clone()

The objects of this class are cloneable with this method.

Usage

CohortSubsetOperator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package CohortGenerator version 0.8.0 Index]