LimitSubsetOperator {CohortGenerator}	R Documentation

Limit Subset Operator

Description

operator to apply limiting subset operations (e.g. washout periods, calendar ranges or earliest entries)

Get auto generated name

Super class

CohortGenerator::SubsetOperator -> LimitSubsetOperator

Active bindings

priorTime

minimum washout time in days

followUpTime

minimum required follow up time in days

limitTo

character one of: "firstEver" - only first entry in patient history "earliestRemaining" - only first entry after washout set by priorTime "latestRemaining" - the latest remaining after washout set by followUpTime "lastEver" - only last entry in patient history inside

Note, when using firstEver and lastEver with follow up and washout, patients with events outside this will be censored.

calendarStartDate

The calendar start date for limiting by date

calendarEndDate

The calendar end date for limiting by date

Methods

Public methods

LimitSubsetOperator$getAutoGeneratedName()

LimitSubsetOperator$toList()

LimitSubsetOperator$clone()

Inherited methods

CohortGenerator::SubsetOperator$classname()

CohortGenerator::SubsetOperator$getQueryBuilder()

CohortGenerator::SubsetOperator$initialize()

CohortGenerator::SubsetOperator$isEqualTo()

CohortGenerator::SubsetOperator$publicFields()

CohortGenerator::SubsetOperator$toJSON()

Method getAutoGeneratedName()

name generated from subset operation properties

Usage

LimitSubsetOperator$getAutoGeneratedName()

Returns

character To List

Method toList()

List representation of object

Usage

LimitSubsetOperator$toList()

Method clone()

The objects of this class are cloneable with this method.

Usage

LimitSubsetOperator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package CohortGenerator version 0.8.0 Index]