createSubsetCohortWindow {CohortGenerator}	R Documentation

A definition of subset functions to be applied to a set of cohorts

Description

A definition of subset functions to be applied to a set of cohorts

Usage

createSubsetCohortWindow(startDay, endDay, targetAnchor)

Arguments

startDay
The start day for the window

endDay
The end day for the window

targetAnchor
To anchor using the target cohort's start date or end date

Value

a SubsetCohortWindow instance

[Package CohortGenerator version 0.8.0 Index]